imported highlighting function from utils
